Extended version of predict.rpart

Identical to predict.rpart but optionally show the node numbers and rules for the predicted values.

rpart.predict(object, newdata,
         type = c("vector", "prob", "class", "matrix"),
         na.action = na.pass,
         nn=FALSE, rules=FALSE, ...)

Arguments

object, newdata, type, na.action
Identical to the same arguments for predict.rpart.
If both nn and rules are FALSE, the returned value is identical to predict.rpart.

nn
If TRUE, return a data.frame with the predictions as usual but with an extra column showing the leaf node number for each prediction.

rules
If TRUE, return a data.frame with the predictions as usual but with an extra column showing the rpart rule (as a string) for each prediction.
It may be helpful to use options(width=1000) before printing this data.frame.

...
Passed on to rpart.rules, for example clip.facs=TRUE.

Value

Same as predict.rpart, but with additional information if nn=TRUE and/or rules=TRUE.

See also

Examples

data(ptitanic)
model <- rpart(survived ~ ., data = ptitanic, cp = .02)
head(rpart.predict(model, rules=TRUE))

#>        died  survived                                                
#> 1 0.2725322 0.7274678 because sex is female                          
#> 2 0.1111111 0.8888889 because sex is   male & age <  9.5 & sibsp <  3
#> 3 0.2725322 0.7274678 because sex is female                          
#> 4 0.8291457 0.1708543 because sex is   male & age >= 9.5             
#> 5 0.2725322 0.7274678 because sex is female                          
#> 6 0.8291457 0.1708543 because sex is   male & age >= 9.5